added appengine yaml files for stocks app so that app can be deployed on app engine flex also added requirements.txt file to each app in Stocks App.
Added a new project Realtime Web Personalisation:
This is a apache beam pipeline that injests web log server data and enriches it with geo ip data and http codes.
This is to practise joins in beam ysing side input.
Currently a batch pipeline but will change to streaming in future commits.

To test locally comment out all GoogleCloudOptions at top of web_data_enrichment.py and set runner to DirectRunner

Tested and working